Update item.rb file by dividing into 3 files
