+++ lib/ead.rb
@@ -1,6 +1,7 @@
 require 'json'
 require 'fileutils'
 require 'item'
+require 'item_clone'
 require 'block'
 require 'byebug'
